+++ packages/editor/src/store/test/reducer.js
@@ -14,6 +14,7 @@ import {
 	saving,
 	postSavingLock,
 	postAutosavingLock,
+	removedPanels,
 } from '../reducer';
 
 describe( 'state', () => {
@@ -262,6 +263,26 @@ describe( 'state', () => {
 			} );
 
 			expect( state ).toEqual( {} );
+		} );
+	} );
+
+	describe( 'removedPanels', () => {
+		it( 'should remove panel', () => {
+			const original = deepFreeze( [] );
+			const state = removedPanels( original, {
+				type: 'REMOVE_PANEL',
+				panelName: 'post-status',
+			} );
+			expect( state ).toEqual( [ 'post-status' ] );
+		} );
+
+		it( 'should not remove already removed panel', () => {
+			const original = deepFreeze( [ 'post-status' ] );
+			const state = removedPanels( original, {
+				type: 'REMOVE_PANEL',
+				panelName: 'post-status',
+			} );
+			expect( state ).toBe( original );
 		} );
 	} );
 } );
